adminPbeta

+++ blogVue/server/routes/index.ts
@@ -3,6 +3,7 @@ import { ObjectId } from 'mongodb'
 import client from '@/db'
 import fs from 'fs'
 const db = client.db('blog', 'posts')
+const dbheader = client.db('blog', 'header');
 const router = Router()
 
 router.get('/posts', async (request, response) => {
@@ -20,14 +21,11 @@ router.put('/posts', async (request, response) => {
   try {
     let data = request.body
     let filedata = request.file;
-    if (!filedata) {
-      console.log("Ошибка при загрузке файла")
-      const res = await db.updateOne({ _id: new ObjectId(data._id) }, { $set: { title: data.title, text: data.text, url: data.url } })
-      response.send({ result: res })
-    } else {
-      const res = await db.updateOne({ _id: new ObjectId(data._id) }, { $set: { title: data.title, text: data.text, url: data.url, src: data.src } })
-      response.send({ result: res })
+    if (Boolean(data.src) && filedata) {
+      fs.unlinkSync(`../client/src/assets/${data.src}`)
     }
+    const res = await db.updateOne({ _id: new ObjectId(data._id) }, { $set: { title: data.title, text: data.text, url: data.url, src: (filedata) ? filedata.originalname : data.src } })
+    response.send({ result: res })
   } catch (e) {
     response.send({ result: 'error', data: e })
   }
@@ -35,13 +33,9 @@ router.put('/posts', async (request, response) => {
 router.post('/posts', async (request, response) => {
   try {
     let filedata = request.file
-    if (!filedata) {
-      console.log("Ошибка при загрузке файла")
-    } else {
-      console.log("Файл загружен")
-      const res = await db.insertOne({ ...request.body, src: filedata.originalname })
-      response.send({ result: res })
-    }
+    const res = await db.insertOne({ ...request.body, src: (filedata) ? filedata.originalname : false })
+    response.send({ result: res })
+
   } catch (e) {
     response.send({ result: 'error', data: e })
   }
@@ -49,7 +43,9 @@ router.post('/posts', async (request, response) => {
 
 router.delete('/posts', async (request, response) => {
   try {
-    fs.unlinkSync(`../client/src/assets/${request.body.p.src}`)
+    if (request.body.p.src) {
+      fs.unlinkSync(`../client/src/assets/${request.body.p.src}`)
+    }
     // fs.unlinkSync(`./public/assets/${request.body.p.src}`)
     const res = await db.deleteOne({ _id: new ObjectId(request.body.p._id) })
     response.send({ result: res })
@@ -59,6 +55,18 @@ router.delete('/posts', async (request, response) => {
   }
 })
 
+router.post('/header', async (request, response) => {
+  try {
+    const res = 'a'
+    const user = await dbheader.findOne({ username: request.body.username });
+    if (!user) {
+      return false;
+    }
+    response.send({ result: res })
+  } catch (e) {
+    response.send({ result: 'error', data: e })
+  }
+})
 
 
 export default router
